requirements.txt
numpy
pyserial
open3d
pyqt5
pyqtgraph
